prevent a known bug to mess up the video order
For some reason, when using a long (>30) list of feeds, videos from a
single feed may fail to properly parse the release time, causing them to
be sorted in front of all other videos. To circumvent this, possible
parsing errors are now assigned an arbitrary high release time to allow
more recent videos to be sorted in front.